Refactor Surface, Tile, Track* actions hpp=>h/cpp
