+++ src/openrct2-ui/windows/Map.cpp
@@ -21,7 +21,7 @@
 #include <openrct2/actions/LandSetRightsAction.h>
 #include <openrct2/actions/PlaceParkEntranceAction.h>
 #include <openrct2/actions/PlacePeepSpawnAction.h>
-#include <openrct2/actions/SurfaceSetStyleAction.hpp>
+#include <openrct2/actions/SurfaceSetStyleAction.h>
 #include <openrct2/audio/audio.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/ride/Track.h>
