+++ src/openrct2-ui/windows/RideConstruction.cpp
@@ -18,9 +18,9 @@
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
 #include <openrct2/actions/RideEntranceExitPlaceAction.h>
-#include <openrct2/actions/TrackPlaceAction.hpp>
-#include <openrct2/actions/TrackRemoveAction.hpp>
-#include <openrct2/actions/TrackSetBrakeSpeedAction.hpp>
+#include <openrct2/actions/TrackPlaceAction.h>
+#include <openrct2/actions/TrackRemoveAction.h>
+#include <openrct2/actions/TrackSetBrakeSpeedAction.h>
 #include <openrct2/audio/audio.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
