+++ src/openrct2-ui/windows/TileInspector.cpp
@@ -14,7 +14,7 @@
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
-#include <openrct2/actions/TileModifyAction.hpp>
+#include <openrct2/actions/TileModifyAction.h>
 #include <openrct2/common.h>
 #include <openrct2/core/Guard.hpp>
 #include <openrct2/localisation/Localisation.h>
