+++ src/openrct2/ReplayManager.cpp
@@ -20,8 +20,8 @@
 #include "actions/RideEntranceExitPlaceAction.h"
 #include "actions/RideSetSettingAction.h"
 #include "actions/SetCheatAction.h"
-#include "actions/TileModifyAction.hpp"
-#include "actions/TrackPlaceAction.hpp"
+#include "actions/TileModifyAction.h"
+#include "actions/TrackPlaceAction.h"
 #include "config/Config.h"
 #include "core/DataSerialiser.h"
 #include "core/Path.hpp"
