+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -76,12 +76,12 @@
 #include "StaffSetNameAction.h"
 #include "StaffSetOrdersAction.h"
 #include "StaffSetPatrolAreaAction.h"
-#include "SurfaceSetStyleAction.hpp"
-#include "TileModifyAction.hpp"
+#include "SurfaceSetStyleAction.h"
+#include "TileModifyAction.h"
 #include "TrackDesignAction.h"
-#include "TrackPlaceAction.hpp"
-#include "TrackRemoveAction.hpp"
-#include "TrackSetBrakeSpeedAction.hpp"
+#include "TrackPlaceAction.h"
+#include "TrackRemoveAction.h"
+#include "TrackSetBrakeSpeedAction.h"
 #include "WallPlaceAction.hpp"
 #include "WallRemoveAction.hpp"
 #include "WallSetColourAction.hpp"
