+++ src/openrct2/ride/TrackDesign.cpp
@@ -25,8 +25,8 @@
 #include "../actions/RideSetVehicleAction.h"
 #include "../actions/SmallSceneryPlaceAction.h"
 #include "../actions/SmallSceneryRemoveAction.h"
-#include "../actions/TrackPlaceAction.hpp"
-#include "../actions/TrackRemoveAction.hpp"
+#include "../actions/TrackPlaceAction.h"
+#include "../actions/TrackRemoveAction.h"
 #include "../actions/WallPlaceAction.hpp"
 #include "../actions/WallRemoveAction.hpp"
 #include "../audio/audio.h"
